refactor: created utils directory and moved Api.js into it
